Everything works, but the task is not fully done
