+++ frontend/src/Views/ProjectView.tsx
@@ -1,36 +1,30 @@
-import { add, archive } from "@equinor/eds-icons"
 import {
-    Button,
-    EdsProvider,
-    Icon,
-    TextField,
-    Tooltip,
-    Typography,
+    Tabs,
 } from "@equinor/eds-core-react"
-import {
-    ChangeEventHandler,
-    MouseEventHandler,
+import React, {
     useEffect,
     useMemo,
     useState,
 } from "react"
-import { useParams, useNavigate } from "react-router-dom"
+import { useParams } from "react-router-dom"
 import styled from "styled-components"
 
 import { Project } from "../models/Project"
 import { GetProjectService } from "../Services/ProjectService"
 
-import { Modal } from "../Components/Modal"
-import { GetCaseService } from "../Services/CaseService"
+import { unwrapProjectId } from "../Utils/common"
 
-import { GetSTEAService } from "../Services/STEAService"
-import { unwrapProjectId, GetProjectCategoryName, GetProjectPhaseName } from "../Utils/common"
-import { WrapperColumn } from "./Asset/StyledAssetComponents"
-import PhysicalUnit from "../Components/PhysicalUnit"
-import Currency from "../Components/Currency"
 import { Case } from "../models/Case"
-import LinearDataTable from "../Components/LinearDataTable"
-import CaseTabs from "./CaseTabs"
+import OverviewView from "./OverviewView"
+import CompareCasesView from "./CompareCasesView"
+
+const { Panel } = Tabs
+const { List, Tab, Panels } = Tabs
+
+const StyledTabPanel = styled(Panel)`
+    padding-top: 0px;
+    border-top: 1px solid LightGray;
+`
 
 const Wrapper = styled.div`
     margin: 2rem;
@@ -38,46 +32,10 @@ const Wrapper = styled.div`
     flex-direction: column;
 `
 
-const Header = styled.header`
-    display: flex;
-    align-items: center;
-
-    > *:first-child {
-        margin-right: 2rem;
-    }
-`
-
-const ProjectDataFieldLabel = styled(Typography)`
-    margin-top: 1rem;
-    font-weight: bold;
-    white-space: pre-wrap;
-`
-
-const ActionsContainer = styled.div`
-    > *:not(:last-child) {
-        margin-right: 0.5rem;
-    }
-`
-
-const ChartsContainer = styled.div`
-    display: flex;
-`
-
-const CreateCaseForm = styled.form`
-    width: 30rem;
-
-    > * {
-        margin-bottom: 1.5rem;
-    }
-`
-
 const ProjectView = () => {
-    const navigate = useNavigate()
+    const [activeTab, setActiveTab] = React.useState(0)
     const params = useParams()
     const [project, setProject] = useState<Project>()
-    const [createCaseModalIsOpen, setCreateCaseModalIsOpen] = useState<boolean>(false)
-    const [caseName, setCaseName] = useState<string>("")
-    const [caseDescription, setCaseDescription] = useState<string>("")
     const [physicalUnit, setPhysicalUnit] = useState<Components.Schemas.PhysUnit>(0)
     const [currency, setCurrency] = useState<Components.Schemas.Currency>(1)
 
@@ -142,158 +100,44 @@ const ProjectView = () => {
         setCapexYearCaseTitles(caseTitles)
     }, [project])
 
-    const toggleCreateCaseModal = () => setCreateCaseModalIsOpen(!createCaseModalIsOpen)
-
-    const handleCaseNameChange: ChangeEventHandler<HTMLInputElement> = (e) => {
-        const { value } = e.target
-        setCaseName(value)
-    }
-
-    const handleDescriptionChange: ChangeEventHandler<HTMLInputElement> = (e) => {
-        const { value } = e.target
-        setCaseDescription(value)
-    }
-
-    const submitToSTEA: MouseEventHandler<HTMLButtonElement> = async (e) => {
-        e.preventDefault()
-
-        try {
-            const projectId: string = unwrapProjectId(params.projectId)
-            const projectResult: Project = await GetProjectService().getProjectByID(projectId)
-            GetSTEAService().excelToSTEA(projectResult)
-        } catch (error) {
-            console.error("[ProjectView] error while submitting form data", error)
-        }
-    }
-
-    const submitCreateCaseForm: MouseEventHandler<HTMLButtonElement> = async (e) => {
-        e.preventDefault()
-
-        try {
-            const projectResult: Project = await GetCaseService().createCase({
-                description: caseDescription,
-                name: caseName,
-                projectId: params.projectId,
-            })
-            toggleCreateCaseModal()
-            navigate(`/project/${projectResult.id}/case/${projectResult.cases.find((o) => (
-                o.name === caseName
-            ))?.id}`)
-        } catch (error) {
-            console.error("[ProjectView] error while submitting form data", error)
-        }
-    }
-
     if (!project) return null
 
     return (
         <Wrapper>
-            <CaseTabs project={project} />
-            <Header>
-                <Typography variant="h2">{project.name}</Typography>
-                <EdsProvider density="compact">
-                    <ActionsContainer>
-                        <Tooltip title="Export to STEA">
-                            <Button
-                                variant="ghost_icon"
-                                aria-label="Export to STEA"
-                                onClick={submitToSTEA}
-                            >
-                                <Icon data={archive} />
-                            </Button>
-                        </Tooltip>
-                        <Tooltip title="Add a case">
-                            <Button variant="ghost_icon" aria-label="Add a case" onClick={toggleCreateCaseModal}>
-                                <Icon data={add} />
-                            </Button>
-                        </Tooltip>
-                    </ActionsContainer>
-                </EdsProvider>
-            </Header>
-            <WrapperColumn>
-                <ProjectDataFieldLabel>Description:</ProjectDataFieldLabel>
-                <Typography variant="h3">{project.description}</Typography>
-            </WrapperColumn>
-            <WrapperColumn>
-                <ProjectDataFieldLabel>Project Phase:</ProjectDataFieldLabel>
-                <Typography variant="h4" aria-label="Project phase">
-                    {GetProjectPhaseName(project.phase)}
-                </Typography>
-            </WrapperColumn>
-            <WrapperColumn>
-                <ProjectDataFieldLabel>Project Category:</ProjectDataFieldLabel>
-                <Typography variant="h4" aria-label="Project category">
-                    {GetProjectCategoryName(project.category)}
-                </Typography>
-            </WrapperColumn>
-            <WrapperColumn>
-                <ProjectDataFieldLabel>Country:</ProjectDataFieldLabel>
-                <Typography variant="h4" aria-label="Country">
-                    {project.country ?? "Not defined in Common Library"}
-                </Typography>
-            </WrapperColumn>
-            <PhysicalUnit
-                currentValue={physicalUnit}
-                setPhysicalUnit={setPhysicalUnit}
-                setProject={setProject}
-                project={project}
-            />
+            <Tabs activeTab={activeTab} onChange={setActiveTab}>
+                <List>
+                    <Tab>Overview </Tab>
+                    <Tab>Compare cases</Tab>
+                    <Tab>Workflow</Tab>
+                    <Tab>Settings</Tab>
+                </List>
+                <Panels>
+                    <StyledTabPanel>
+                        <OverviewView
+                            project={project}
+                            setProject={setProject}
+                            physicalUnit={physicalUnit}
+                            setPhysicalUnit={setPhysicalUnit}
+                            currency={currency}
+                            setCurrency={setCurrency}
+                        />
+                    </StyledTabPanel>
+                    <StyledTabPanel>
 
-            <Currency
-                currentValue={currency}
-                setCurrency={setCurrency}
-                setProject={setProject}
-                project={project}
-            />
-            <ChartsContainer>
-                {capexYearXLabels.length !== 0
-                    ? (
-                        <LinearDataTable
+                        <CompareCasesView
                             capexYearX={capexYearXLabels}
                             capexYearY={capexYearYDatas}
                             caseTitles={capexYearCaseTitles}
                         />
-                    )
-                    : <Typography> No cases containing CapEx to display data for</Typography> }
-            </ChartsContainer>
-
-            <Modal isOpen={createCaseModalIsOpen} title="Create a case" shards={[]}>
-                <CreateCaseForm>
-                    <TextField
-                        label="Name"
-                        id="name"
-                        name="name"
-                        placeholder="Enter a name"
-                        onChange={handleCaseNameChange}
-                    />
-
-                    <TextField
-                        label="Description"
-                        id="description"
-                        name="description"
-                        placeholder="Enter a description"
-                        onChange={handleDescriptionChange}
-                    />
-
-                    <div>
-                        <Button
-                            type="submit"
-                            onClick={submitCreateCaseForm}
-                            disabled={caseName === "" || caseDescription === ""}
-                        >
-                            Create case
-                        </Button>
-                        <Button
-                            type="button"
-                            color="secondary"
-                            variant="ghost"
-                            onClick={toggleCreateCaseModal}
-                        >
-                            Cancel
-                        </Button>
-                    </div>
-                </CreateCaseForm>
-            </Modal>
+                    </StyledTabPanel>
+                    <StyledTabPanel>
+                        <p>Workflow</p>
+                    </StyledTabPanel>
+                    <StyledTabPanel>
+                        <p>Setting</p>
+                    </StyledTabPanel>
+                </Panels>
+            </Tabs>
         </Wrapper>
     )
 }
